:recycle: 简化关于页更新状态订阅与版本检查流程
统一两端单击检查入口，直接订阅状态，合并版本比较函数。iOS 仅由关于页触发自动检查，保留必要的本地版本平台实现、共享请求和 Debug 预览隔离。
