+++ cyxbs-pages/mine/src/commonMain/kotlin/com/cyxbs/pages/mine/about/ui/AboutNavEntry.kt
@@ -20,6 +20,7 @@ import androidx.compose.material.Text
 import androidx.compose.material.TopAppBar
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
@@ -34,6 +35,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.constraintlayout.compose.ConstraintLayout
 import androidx.constraintlayout.compose.ConstraintSet
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.cyxbs.components.config.APP_WEBSITE
 import com.cyxbs.components.config.ICP_WEBSITE
 import com.cyxbs.components.config.Platform
@@ -59,7 +61,6 @@ import com.cyxbs.functions.update.api.IAppUpdateService
 import com.cyxbs.pages.login.api.ILegalNoticeService
 import cyxbsmobile.cyxbs_pages.mine.generated.resources.Res
 import cyxbsmobile.cyxbs_pages.mine.generated.resources.mine_ic_arrow_right
-import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.launch
 import kotlinx.datetime.TimeZone
 import kotlinx.datetime.toLocalDateTime
@@ -80,312 +81,308 @@ object AboutNavArgument : AppNavArgument
 @AppNav(route = NAV_ABOUT)
 class AboutNavEntry : AppNavEntry<AboutNavArgument>() {
 
-    override fun isNeedLogin(argument: AboutNavArgument): Boolean {
-        return false
-    }
+  override fun isNeedLogin(argument: AboutNavArgument): Boolean {
+    return false
+  }
 
-    @Composable
-    override fun Content(argument: AboutNavArgument) {
-        AboutPage(argument)
-    }
+  @Composable
+  override fun Content(argument: AboutNavArgument) {
+    AboutPage(argument)
+  }
 }
 
 @Composable
 private fun AboutPage(argument: AboutNavArgument) {
-    ConstraintLayout(
-        constraintSet = createConstraintSet(),
-        modifier = Modifier.fillMaxSize()
-            .background(LocalAppColors.current.bottomBg)
-            .statusBarsPadding(),
-        animateChangesSpec = spring(
-            stiffness = Spring.StiffnessMediumLow
-        )
-    ) {
-        TopBarCompose(argument = argument, modifier = Modifier.layoutId(Element.Topbar))
-        LogoCompose(modifier = Modifier.layoutId(Element.Logo))
-        AppInfoCompose(modifier = Modifier.layoutId(Element.AppInfo))
-        BackgroundIvCompose(modifier = Modifier.layoutId(Element.BackgroundIv))
-        VersionUpdateCompose(modifier = Modifier.layoutId(Element.VersionUpdate))
-        VersionInfoCompose(modifier = Modifier.layoutId(Element.VersionInfo))
-        ProductWebsiteCompose(modifier = Modifier.layoutId(Element.ProductWebsite))
-        ShareCompose(modifier = Modifier.layoutId(Element.Share))
-        BottomInfoCompose(modifier = Modifier.navigationBarsPadding().layoutId(Element.BottomInfo))
-    }
+  ConstraintLayout(
+    constraintSet = createConstraintSet(),
+    modifier = Modifier.fillMaxSize()
+      .background(LocalAppColors.current.bottomBg)
+      .statusBarsPadding(),
+    animateChangesSpec = spring(
+      stiffness = Spring.StiffnessMediumLow
+    )
+  ) {
+    TopBarCompose(argument = argument, modifier = Modifier.layoutId(Element.Topbar))
+    LogoCompose(modifier = Modifier.layoutId(Element.Logo))
+    AppInfoCompose(modifier = Modifier.layoutId(Element.AppInfo))
+    BackgroundIvCompose(modifier = Modifier.layoutId(Element.BackgroundIv))
+    VersionUpdateCompose(modifier = Modifier.layoutId(Element.VersionUpdate))
+    VersionInfoCompose(modifier = Modifier.layoutId(Element.VersionInfo))
+    ProductWebsiteCompose(modifier = Modifier.layoutId(Element.ProductWebsite))
+    ShareCompose(modifier = Modifier.layoutId(Element.Share))
+    BottomInfoCompose(modifier = Modifier.navigationBarsPadding().layoutId(Element.BottomInfo))
+  }
 }
 
 @Composable
 private fun createConstraintSet(): ConstraintSet {
-    val windowSize = getWindowScreenSize()
-    return ConstraintSet {
-        AboutConstraintSet(
-            scope = this,
-            windowSize = windowSize
-        ).createConstrain()
-    }
+  val windowSize = getWindowScreenSize()
+  return ConstraintSet {
+    AboutConstraintSet(
+      scope = this,
+      windowSize = windowSize
+    ).createConstrain()
+  }
 }
 
 @Composable
 private fun TopBarCompose(argument: AboutNavArgument, modifier: Modifier = Modifier) {
-    TopAppBar(
-        modifier = modifier,
-        backgroundColor = Color.Transparent,
-        elevation = 0.dp
+  TopAppBar(
+    modifier = modifier,
+    backgroundColor = Color.Transparent,
+    elevation = 0.dp
+  ) {
+    Box(
+      modifier = Modifier.fillMaxSize()
     ) {
-        Box(
-            modifier = Modifier.fillMaxSize()
-        ) {
-            Image(
-                painter = painterResource(ConfigRes.configIcBack()),
-                contentDescription = null,
-                modifier = Modifier
-                    .align(Alignment.CenterStart)
-                    .padding(start = 20.dp)
-                    .size(16.dp)
-                    .clickableSingle {
-                        argument.popBackStack()
-                    }
-            )
-            Text(
-                text = "关于我们",
-                fontSize = 20.sp,
-                color = LocalAppColors.current.tvLv1,
-                fontWeight = FontWeight.SemiBold,
-                modifier = Modifier.align(Alignment.Center),
-                textAlign = TextAlign.Center
-            )
-        }
+      Image(
+        painter = painterResource(ConfigRes.configIcBack()),
+        contentDescription = null,
+        modifier = Modifier
+          .align(Alignment.CenterStart)
+          .padding(start = 20.dp)
+          .size(16.dp)
+          .clickableSingle {
+            argument.popBackStack()
+          }
+      )
+      Text(
+        text = "关于我们",
+        fontSize = 20.sp,
+        color = LocalAppColors.current.tvLv1,
+        fontWeight = FontWeight.SemiBold,
+        modifier = Modifier.align(Alignment.Center),
+        textAlign = TextAlign.Center
+      )
     }
+  }
 }
 
 @Composable
 private fun LogoCompose(modifier: Modifier = Modifier) {
-    Image(
-        modifier = modifier.size(96.dp, 96.dp),
-        painter = painterResource(ConfigRes.configIcAppLogo()),
-        contentDescription = null
-    )
+  Image(
+    modifier = modifier.size(96.dp, 96.dp),
+    painter = painterResource(ConfigRes.configIcAppLogo()),
+    contentDescription = null
+  )
 }
 
 @Composable
 private fun AppInfoCompose(modifier: Modifier = Modifier) {
-    Column(
-        modifier = modifier,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Text(
-            text = "掌上重邮",
-            fontSize = 20.sp,
-            color = LocalAppColors.current.tvLv2,
-            fontWeight = FontWeight.Bold
-        )
-        Text(
-            modifier = Modifier.padding(top = 5.dp),
-            text = "Version ${getAppVersionName()}",
-            fontSize = 13.sp,
-            color = LocalAppColors.current.tvLv2
-        )
-    }
+  Column(
+    modifier = modifier,
+    horizontalAlignment = Alignment.CenterHorizontally
+  ) {
+    Text(
+      text = "掌上重邮",
+      fontSize = 20.sp,
+      color = LocalAppColors.current.tvLv2,
+      fontWeight = FontWeight.Bold
+    )
+    Text(
+      modifier = Modifier.padding(top = 5.dp),
+      text = "Version ${getAppVersionName()}",
+      fontSize = 13.sp,
+      color = LocalAppColors.current.tvLv2
+    )
+  }
 }
 
 @Composable
 private fun BackgroundIvCompose(modifier: Modifier = Modifier) {
-    Box(
-        modifier = modifier
-            .background(
-                color = LocalAppColors.current.topBg,
-                shape = RoundedCornerShape(topStart = 16.dp, topEnd = 16.dp)
-            )
-    )
+  Box(
+    modifier = modifier
+      .background(
+        color = LocalAppColors.current.topBg,
+        shape = RoundedCornerShape(topStart = 16.dp, topEnd = 16.dp)
+      )
+  )
 }
 
 @Composable
 private fun VersionUpdateCompose(modifier: Modifier = Modifier) {
-    val enableDebugUpdate = isDebug()
-    val updateStatus = remember { mutableStateOf("") }
-    val manualCheckInProgress = remember { mutableStateOf(false) }
-    val coroutineScope = rememberCoroutineScope()
-    Box(
-        modifier = modifier
-            .combinedClickable(
-                interactionSource = remember { MutableInteractionSource() },
-                indication = DefaultIndication,
-                onClick = {
-                    val appUpdateService = IAppUpdateService
-                    val nowUpdateStatus = appUpdateService.getUpdateStatus().value
-                    if (appPlatform != Platform.IOS && nowUpdateStatus is AppUpdateStatus.Result.Dated) {
-                        appUpdateService.noticeUpdate(nowUpdateStatus.newVersion)
-                    } else if (!manualCheckInProgress.value) {
-                        manualCheckInProgress.value = true
-                        coroutineScope.launch {
-                            try {
-                                when (val status = appUpdateService.checkUpdate()) {
-                                    AppUpdateStatus.Result.Valid -> toast("已经是最新版了哦")
-                                    is AppUpdateStatus.Result.Dated -> appUpdateService.noticeUpdate(status.newVersion)
-                                    is AppUpdateStatus.Result.Error -> {
-                                        toast(
-                                            if (appPlatform == Platform.IOS) "检查更新失败，请稍后重试"
-                                            else "有一股神秘力量阻拦了更新，可尝试点击「产品官网」下载最新版～"
-                                        )
-                                    }
-                                }
-                            } finally {
-                                manualCheckInProgress.value = false
-                            }
-                        }
-                    }
-                },
-                onLongClick = {
-                    if (enableDebugUpdate) {
-                        IAppUpdateService::class.implOrNull()?.debug()
-                    }
+  val enableDebugUpdate = isDebug()
+  val appUpdateService = IAppUpdateService
+  val updateStatus by appUpdateService.getUpdateStatus().collectAsStateWithLifecycle()
+  val manualCheckInProgress = remember { mutableStateOf(false) }
+  val coroutineScope = rememberCoroutineScope()
+  Box(
+    modifier = modifier
+      .combinedClickable(
+        interactionSource = remember { MutableInteractionSource() },
+        indication = DefaultIndication,
+        onClick = {
+          if (!manualCheckInProgress.value) {
+            manualCheckInProgress.value = true
+            coroutineScope.launch {
+              try {
+                when (val status = appUpdateService.checkUpdate()) {
+                  AppUpdateStatus.Result.Valid -> toast("已经是最新版了哦")
+                  is AppUpdateStatus.Result.Dated -> appUpdateService.noticeUpdate(status.newVersion)
+                  is AppUpdateStatus.Result.Error -> {
+                    toast(
+                      if (appPlatform == Platform.IOS) "检查更新失败，请稍后重试"
+                      else "有一股神秘力量阻拦了更新，可尝试点击「产品官网」下载最新版～"
+                    )
+                  }
                 }
-            )
-            .padding(top = 9.dp, bottom = 9.dp)
-    ) {
-        Text(
-            modifier = Modifier.padding(start = 20.dp),
-            text = "版本更新" + if (enableDebugUpdate) " (长按测试)" else "",
-            color = LocalAppColors.current.tvLv2,
-            fontSize = 16.sp
-        )
-        Text(
-            modifier = Modifier.align(Alignment.CenterEnd).padding(end = 38.dp),
-            text = updateStatus.value,
-            fontSize = 13.sp,
-            color = 0x80294169.dark(0x48F0F0F2)
-        )
-        Image(
-            modifier = Modifier.padding(end = 20.dp).size(width = 6.dp, height = 13.dp)
-                .align(Alignment.CenterEnd),
-            painter = painterResource(Res.drawable.mine_ic_arrow_right),
-            contentDescription = null
-        )
-    }
-    LaunchedEffect(Unit) {
-        if (appPlatform == Platform.IOS) {
-            // 进入关于页只刷新状态；更新弹窗由用户手动检查触发。
-            launch { IAppUpdateService.checkUpdate() }
-        }
-        IAppUpdateService.getUpdateStatus().collectLatest {
-            updateStatus.value = when (it) {
-                AppUpdateStatus.Checking -> "检查中..."
-                AppUpdateStatus.Result.Valid -> "已是最新版本"
-                is AppUpdateStatus.Result.Dated -> "发现新版本"
-                is AppUpdateStatus.Result.Error -> "新版本请求失败 >_<"
+              } finally {
+                manualCheckInProgress.value = false
+              }
             }
+          }
+        },
+        onLongClick = {
+          if (enableDebugUpdate) {
+            appUpdateService.debug()
+          }
         }
+      )
+      .padding(top = 9.dp, bottom = 9.dp)
+  ) {
+    Text(
+      modifier = Modifier.padding(start = 20.dp),
+      text = "版本更新" + if (enableDebugUpdate) " (长按测试)" else "",
+      color = LocalAppColors.current.tvLv2,
+      fontSize = 16.sp
+    )
+    Text(
+      modifier = Modifier.align(Alignment.CenterEnd).padding(end = 38.dp),
+      text = when (updateStatus) {
+        AppUpdateStatus.Checking -> "检查中..."
+        AppUpdateStatus.Result.Valid -> "已是最新版本"
+        is AppUpdateStatus.Result.Dated -> "发现新版本"
+        is AppUpdateStatus.Result.Error -> "新版本请求失败 >_<"
+      },
+      fontSize = 13.sp,
+      color = 0x80294169.dark(0x48F0F0F2)
+    )
+    Image(
+      modifier = Modifier.padding(end = 20.dp).size(width = 6.dp, height = 13.dp)
+        .align(Alignment.CenterEnd),
+      painter = painterResource(Res.drawable.mine_ic_arrow_right),
+      contentDescription = null
+    )
+  }
+  LaunchedEffect(appUpdateService) {
+    if (appPlatform == Platform.IOS) {
+      // 进入关于页只刷新状态；更新弹窗由用户手动检查触发。
+      appUpdateService.checkUpdate()
     }
+  }
 }
 
 @Composable
 private fun VersionInfoCompose(modifier: Modifier = Modifier) {
-    val showState = remember { mutableStateOf(false) }
-    InfoItem(modifier, "版本信息") {
-        showState.value = true
-    }
-    UpdateInfoDialog(showState)
+  val showState = remember { mutableStateOf(false) }
+  InfoItem(modifier, "版本信息") {
+    showState.value = true
+  }
+  UpdateInfoDialog(showState)
 }
 
 @Composable
 private fun ProductWebsiteCompose(modifier: Modifier = Modifier) {
-    val uriHandler = LocalUriHandler.current
-    InfoItem(modifier, "产品官网") {
-        uriHandler.openUri(APP_WEBSITE)
-    }
+  val uriHandler = LocalUriHandler.current
+  InfoItem(modifier, "产品官网") {
+    uriHandler.openUri(APP_WEBSITE)
+  }
 }
 
 @Composable
 private fun ShareCompose(modifier: Modifier = Modifier) {
-    InfoItem(modifier, "分享") {
-        shareText("掌上重邮是重邮首款校园生活类App，拥有查课表，签到，邮问等功能，记得分享给好友哦。 $APP_WEBSITE")
-    }
+  InfoItem(modifier, "分享") {
+    shareText("掌上重邮是重邮首款校园生活类App，拥有查课表，签到，邮问等功能，记得分享给好友哦。 $APP_WEBSITE")
+  }
 }
 
 @Composable
 private fun BottomInfoCompose(modifier: Modifier = Modifier) {
-    val uriHandler = LocalUriHandler.current
-    val blueTextColor = 0xFF0BCCF0.dark(0xFF0BCCF0)
-    val declareTextColor = 0x64294169.dark(0x48F0F0F2)
-    val currentYear = Clock.System.now().toLocalDateTime(TimeZone.currentSystemDefault()).year
-    Column(
-        modifier = modifier,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Row(modifier = Modifier.padding(bottom = 8.dp)) {
-            Text(
-                text = "《用户协议》",
-                fontSize = 11.sp,
-                color = blueTextColor,
-                modifier = Modifier.clickableNoIndicator {
-                    ILegalNoticeService::class.implOrNull()?.openUserAgreementScreen() ?: toast("当前应用未实现「用户协议」功能")
-                }
-            )
-            Text(
-                text = "&",
-                fontSize = 11.sp,
-                color = LocalAppColors.current.tvLv2
-            )
-            Text(
-                text = "《隐私政策》",
-                fontSize = 11.sp,
-                color = blueTextColor,
-                modifier = Modifier.clickableSingle {
-                    ILegalNoticeService::class.implOrNull()?.openPrivacyPolicyScreen() ?: toast("当前应用未实现「隐私政策」功能")
-                }
-            )
+  val uriHandler = LocalUriHandler.current
+  val blueTextColor = 0xFF0BCCF0.dark(0xFF0BCCF0)
+  val declareTextColor = 0x64294169.dark(0x48F0F0F2)
+  val currentYear = Clock.System.now().toLocalDateTime(TimeZone.currentSystemDefault()).year
+  Column(
+    modifier = modifier,
+    horizontalAlignment = Alignment.CenterHorizontally
+  ) {
+    Row(modifier = Modifier.padding(bottom = 8.dp)) {
+      Text(
+        text = "《用户协议》",
+        fontSize = 11.sp,
+        color = blueTextColor,
+        modifier = Modifier.clickableNoIndicator {
+          ILegalNoticeService::class.implOrNull()?.openUserAgreementScreen()
+            ?: toast("当前应用未实现「用户协议」功能")
         }
-        Text(
-            modifier = Modifier.padding(top = 5.dp),
-            text = "红岩网校工作站出品",
-            fontSize = 11.sp,
-            color = declareTextColor
-        )
-        Text(
-            modifier = Modifier.padding(top = 5.dp),
-            text = "CopyRight © 2015-${currentYear} All Rights Reserved",
-            fontSize = 11.sp,
-            color = declareTextColor
-        )
-        Text(
-            modifier = Modifier
-                .padding(top = 5.dp)
-                .clickableSingle {
-                    uriHandler.openUri(ICP_WEBSITE)
-                },
-            text = "ICP备案号: 渝ICP备17002788号-7A",
-            fontSize = 11.sp,
-            color = blueTextColor
-        )
+      )
+      Text(
+        text = "&",
+        fontSize = 11.sp,
+        color = LocalAppColors.current.tvLv2
+      )
+      Text(
+        text = "《隐私政策》",
+        fontSize = 11.sp,
+        color = blueTextColor,
+        modifier = Modifier.clickableSingle {
+          ILegalNoticeService::class.implOrNull()?.openPrivacyPolicyScreen()
+            ?: toast("当前应用未实现「隐私政策」功能")
+        }
+      )
     }
+    Text(
+      modifier = Modifier.padding(top = 5.dp),
+      text = "红岩网校工作站出品",
+      fontSize = 11.sp,
+      color = declareTextColor
+    )
+    Text(
+      modifier = Modifier.padding(top = 5.dp),
+      text = "CopyRight © 2015-${currentYear} All Rights Reserved",
+      fontSize = 11.sp,
+      color = declareTextColor
+    )
+    Text(
+      modifier = Modifier
+        .padding(top = 5.dp)
+        .clickableSingle {
+          uriHandler.openUri(ICP_WEBSITE)
+        },
+      text = "ICP备案号: 渝ICP备17002788号-7A",
+      fontSize = 11.sp,
+      color = blueTextColor
+    )
+  }
 }
 
 @Composable
 private fun InfoItem(
-    modifier: Modifier = Modifier,
-    text: String,
-    onClick: () -> Unit
+  modifier: Modifier = Modifier,
+  text: String,
+  onClick: () -> Unit
 ) {
-    val interactionSource = MutableInteractionSource()
-    Box(
-        modifier = modifier
-            .clickable(
-                interactionSource = interactionSource,
-                indication = DefaultIndication,
-                onClick = onClick
-            )
-            .padding(top = 9.dp, bottom = 9.dp)
-    ) {
-        Text(
-            modifier = Modifier.padding(start = 20.dp),
-            text = text,
-            color = LocalAppColors.current.tvLv2,
-            fontSize = 16.sp
-        )
-        Image(
-            modifier = Modifier.padding(end = 20.dp).size(width = 6.dp, height = 13.dp)
-                .align(Alignment.CenterEnd),
-            painter = painterResource(Res.drawable.mine_ic_arrow_right),
-            contentDescription = null
-        )
-    }
+  val interactionSource = MutableInteractionSource()
+  Box(
+    modifier = modifier
+      .clickable(
+        interactionSource = interactionSource,
+        indication = DefaultIndication,
+        onClick = onClick
+      )
+      .padding(top = 9.dp, bottom = 9.dp)
+  ) {
+    Text(
+      modifier = Modifier.padding(start = 20.dp),
+      text = text,
+      color = LocalAppColors.current.tvLv2,
+      fontSize = 16.sp
+    )
+    Image(
+      modifier = Modifier.padding(end = 20.dp).size(width = 6.dp, height = 13.dp)
+        .align(Alignment.CenterEnd),
+      painter = painterResource(Res.drawable.mine_ic_arrow_right),
+      contentDescription = null
+    )
+  }
 }
